server/requirements.txt
fastapi[all]
fastapi-socketio
python-socketio